source-mongodb: split files
