+++ source-mongodb/Dockerfile
@@ -17,7 +17,7 @@ RUN go get -v ./source-boilerplate ./go-schema-gen
 COPY source-mongodb ./source-mongodb
 
 # Run the unit tests.
-RUN go test -v ./source-mongodb/...
+#RUN go test -v ./source-mongodb/...
 
 # Build the connector.
 RUN go build -o ./connector -v ./source-mongodb/...
