+++ source-mongodb/main.go
@@ -4,19 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"math"
-	"errors"
-	"io"
 	"net/url"
 
 	schemagen "github.com/estuary/connectors/go-schema-gen"
 	boilerplate "github.com/estuary/connectors/source-boilerplate"
 	pc "github.com/estuary/flow/go/protocols/capture"
 	pf "github.com/estuary/flow/go/protocols/flow"
-	"golang.org/x/sync/errgroup"
-	log "github.com/sirupsen/logrus"
-
-	"github.com/invopop/jsonschema"
 
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/bson"
@@ -126,79 +119,6 @@ func (driver) Spec(ctx context.Context, req *pc.SpecRequest) (*pc.SpecResponse, 
 	}, nil
 }
 
-// minimalSchema is the maximally-permissive schema which just specifies the
-// _id key. The schema of collections is minimalSchema as we
-// rely on Flow's schema inference to infer the collection schema
-var minimalSchema = generateMinimalSchema()
-const idProperty = "_id"
-
-func generateMinimalSchema() json.RawMessage {
-	// Wrap metadata into an enclosing object schema with a /_meta property
-	// and a 'maximize by timestamp' reduction strategy.
-	var schema = &jsonschema.Schema{
-		Type:                 "object",
-		Required:             []string{idProperty},
-		AdditionalProperties: nil,
-		Extras: map[string]interface{}{
-			"properties": map[string]*jsonschema.Schema{
-				idProperty: &jsonschema.Schema{
-					Type: "string",
-				},
-			},
-			"x-infer-schema": true,
-		},
-	}
-
-	// Marshal schema to JSON
-	bs, err := json.Marshal(schema)
-	if err != nil {
-		panic(fmt.Errorf("error generating schema: %v", err))
-	}
-	return json.RawMessage(bs)
-}
-
-// Discover returns the set of resources available from this Driver.
-func (d *driver) Discover(ctx context.Context, req *pc.DiscoverRequest) (*pc.DiscoverResponse, error) {
-	var cfg config
-	if err := pf.UnmarshalStrict(req.EndpointSpecJson, &cfg); err != nil {
-		return nil, fmt.Errorf("parsing config json: %w", err)
-	}
-
-	var client, err = d.Connect(ctx, cfg)
-	if err != nil {
-		return nil, fmt.Errorf("connecting to database: %w", err)
-	}
-
-	defer func() {
-		if err = client.Disconnect(ctx); err != nil {
-			panic(err)
-		}
-	}()
-
-	var db = client.Database(cfg.Database)
-
-	collections, err := db.ListCollectionSpecifications(ctx, bson.D{})
-	if err != nil {
-		return nil, fmt.Errorf("listing collections: %w", err)
-	}
-
-	var bindings = []*pc.DiscoverResponse_Binding{}
-	for _, collection := range collections {
-		resourceJSON, err := json.Marshal(resource{Database: db.Name(), Collection: collection.Name})
-		if err != nil {
-			return nil, fmt.Errorf("serializing resource json: %w", err)
-		}
-
-		bindings = append(bindings, &pc.DiscoverResponse_Binding{
-			RecommendedName: pf.Collection(fmt.Sprintf("%s/%s", db.Name(), collection.Name)),
-			ResourceSpecJson: resourceJSON,
-			DocumentSchemaJson: minimalSchema,
-			KeyPtrs: []string{"/" + idProperty},
-		})
-	}
-
-	return &pc.DiscoverResponse{Bindings: bindings}, nil
-}
 
 // Validate that store resources and proposed collection bindings are compatible.
 func (d *driver) Validate(ctx context.Context, req *pc.ValidateRequest) (*pc.ValidateResponse, error) {
@@ -253,186 +173,6 @@ func (d *driver) ApplyUpsert(ctx context.Context, req *pc.ApplyRequest) (*pc.App
 // ApplyDelete deletes an existing capture from the store.
 func (driver) ApplyDelete(ctx context.Context, req *pc.ApplyRequest) (*pc.ApplyResponse, error) {
 	return &pc.ApplyResponse{ActionDescription: ""}, nil
-}
-
-// Pull is a very long lived RPC through which the Flow runtime and a
-// Driver cooperatively execute an unbounded number of transactions.
-func (d *driver) Pull(stream pc.Driver_PullServer) error {
-	var open, err = stream.Recv()
-	if err != nil {
-		return fmt.Errorf("error reading PullRequest: %w", err)
-	} else if open.Open == nil {
-		return fmt.Errorf("expected PullRequest.Open, got %#v", open)
-	}
-
-	var cfg config
-	if err := pf.UnmarshalStrict(open.Open.Capture.EndpointSpecJson, &cfg); err != nil {
-		return fmt.Errorf("parsing config json: %w", err)
-	}
-
-	var ctx = context.Background()
-
-	client, err := d.Connect(ctx, cfg)
-	if err != nil {
-		return fmt.Errorf("connecting to database: %w", err)
-	}
-	log.Info("connected to database")
-
-	defer func() {
-		if err = client.Disconnect(ctx); err != nil {
-			panic(err)
-		}
-	}()
-
-	var c = capture{
-		Output: &boilerplate.PullOutput{Stream: stream},
-	}
-
-	eg, ctx := errgroup.WithContext(ctx)
-
-	if err := c.Output.Ready(); err != nil {
-		return err
-	}
-
-	for idx, binding := range open.Open.Capture.Bindings {
-		var res resource
-		if err := pf.UnmarshalStrict(binding.ResourceSpecJson, &res); err != nil {
-			return fmt.Errorf("parsing resource config: %w", err)
-		}
-
-		eg.Go(func() error { return c.BackfillCollection(ctx, client, uint32(idx), res) })
-		eg.Go(func() error { return c.ChangeStream(ctx, client, uint32(idx), res) })
-	}
-
-	if err := eg.Wait(); err != nil && !errors.Is(err, io.EOF) {
-		return err
-	}
-
-	return nil
-}
-
-type changeEvent struct{
-	OperationType string `bson:"operationType"`
-	FullDocument bson.M `bson:"fullDocument"`
-}
-
-func (c *capture) ChangeStream(ctx context.Context, client *mongo.Client, binding uint32, res resource) error {
-	var db = client.Database(res.Database)
-
-	var collection = db.Collection(res.Collection)
-
-	log.Debug("listening on changes on collection ", res.Collection)
-	var eventFilter = bson.D{{"$match", bson.D{{"$or",
-	bson.A{
-		// bson.D{{"operationType", "delete"}}, TODO: how do we handle deletion
-		// events
-		bson.D{{"operationType", "insert"}},
-		bson.D{{"operationType", "update"}},
-	}}}}}
-	cursor, err := collection.Watch(ctx, mongo.Pipeline{eventFilter}, options.ChangeStream().SetFullDocument(options.UpdateLookup))
-	if err != nil {
-		return fmt.Errorf("change stream on collection %s: %w", res.Collection, err)
-	}
-
-	for cursor.Next(ctx) {
-		var ev changeEvent
-		if err = cursor.Decode(&ev); err != nil {
-			return fmt.Errorf("decoding document in collection %s: %w", res.Collection, err)
-		}
-
-		var doc = sanitizeDocument(ev.FullDocument)
-
-		js, err := json.Marshal(doc)
-		if err != nil {
-			return fmt.Errorf("encoding document in collection %s as json: %w", res.Collection, err)
-		}
-
-		if err = c.Output.Documents(binding, js); err != nil {
-			return fmt.Errorf("output documents failed: %w", err)
-		}
-
-		if err = c.Output.Checkpoint([]byte("{}"), true); err != nil {
-			return fmt.Errorf("output checkpoint failed: %w", err)
-		}
-	}
-
-	if err := cursor.Err(); err != nil {
-		return fmt.Errorf("cursor error for change stream %s: %w", res.Collection, err)
-	}
-
-	defer cursor.Close(ctx)
-
-	return nil
-}
-
-const CHECKPOINT_EVERY = 100
-
-func (c *capture) BackfillCollection(ctx context.Context, client *mongo.Client, binding uint32, res resource) error {
-	var db = client.Database(res.Database)
-
-	var collection = db.Collection(res.Collection)
-
-	log.Debug("finding documents in collection ", res.Collection)
-	cursor, err := collection.Find(ctx, bson.D{})
-	if err != nil {
-		return fmt.Errorf("could not run find query on collection %s: %w", res.Collection, err)
-	}
-
-	defer cursor.Close(ctx)
-
-	var i = 0
-	for cursor.Next(ctx) {
-		i += 1
-		var doc bson.M
-		if err = cursor.Decode(&doc); err != nil {
-			return fmt.Errorf("decoding document in collection %s: %w", res.Collection, err)
-		}
-		doc = sanitizeDocument(doc)
-
-		js, err := json.Marshal(doc)
-		if err != nil {
-			return fmt.Errorf("encoding document in collection %s as json: %w", res.Collection, err)
-		}
-
-		if err = c.Output.Documents(binding, js); err != nil {
-			return fmt.Errorf("output documents failed: %w", err)
-		}
-
-		if i % CHECKPOINT_EVERY == 0 {
-			if err = c.Output.Checkpoint([]byte("{}"), true); err != nil {
-				return fmt.Errorf("output checkpoint failed: %w", err)
-			}
-		}
-	}
-
-	if err = c.Output.Checkpoint([]byte("{}"), true); err != nil {
-		return fmt.Errorf("sending checkpoint failed: %w", err)
-	}
-
-	if err := cursor.Err(); err != nil {
-		return fmt.Errorf("cursor error for backfill %s: %w", res.Collection, err)
-	}
-
-	return nil
-}
-
-func sanitizeDocument(doc map[string]interface{}) map[string]interface{} {
-	for key, value := range doc {
-		switch v := value.(type) {
-		case float64:
-			if math.IsNaN(v) {
-				doc[key] = "NaN"
-			}
-		case map[string]interface{}:
-			doc[key] = sanitizeDocument(v)
-		}
-	}
-
-	return doc
-}
-
-type capture struct{
-	Output   *boilerplate.PullOutput
 }
 
 func main() {
